implement basic features of plugin API
